# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bridgepath Network API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/{userId}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProfile
      tags: [profiles]
      summary: Update a user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
...
  /payments/create-checkout:
    post:
      operationId: createCheckoutSession
      tags: [payments]
      summary: Create Stripe checkout session for human review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutBody"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify a payment session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentBody"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        bio:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        experience:
          type: string
          nullable: true
        education:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        portfolioUrl:
          type: string
          nullable: true
        resumeUrl:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        companyWebsite:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        companySize:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - skills
        - createdAt
        - updatedAt
...
    UpdateProfileBody:
      type: object
      properties:
        bio:
          type: string
        location:
          type: string
        country:
          type: string
        skills:
          type: array
          items:
            type: string
        experience:
          type: string
        education:
          type: string
        linkedinUrl:
          type: string
        portfolioUrl:
          type: string
        resumeUrl:
          type: string
        companyName:
          type: string
        companyWebsite:
          type: string
        industry:
          type: string
        companySize:
          type: string
...
    CreateCheckoutBody:
      type: object
      properties:
        reviewId:
          type: integer
        successUrl:
          type: string
        cancelUrl:
          type: string
      required:
        - reviewId
        - successUrl
        - cancelUrl
    CheckoutResponse:
      type: object
      properties:
        sessionId:
          type: string
        url:
          type: string
      required:
        - sessionId
        - url
    VerifyPaymentBody:
      type: object
      properties:
        sessionId:
          type: string
      required:
        - sessionId
    PaymentVerifyResponse:
      type: object
      properties:
        paid:
          type: boolean
        reviewId:
          type: integer
          nullable: true
      required:
        - paid
...